cookbook/integrations/flytekit_plugins/xgboost_example/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
attrs==21.2.0
    # via scantree
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.7
    # via requests
click==7.1.2
    # via flytekit
croniter==1.0.15
    # via flytekit
cycler==0.11.0
    # via matplotlib
dataclasses-json==0.5.6
    # via flytekit
decorator==5.1.0
    # via retry
deprecated==1.2.13
    # via flytekit
dirhash==0.2.1
    # via flytekit
diskcache==5.2.1
    # via flytekit
docker-image-py==0.1.12
    # via flytekit
docstring-parser==0.12
    # via flytekit
flyteidl==0.21.8
    # via flytekit
flytekit==0.23.1
    # via
    #   -r ../../../common/requirements-common.in
    #   flytekitplugins-xgboost
flytekitplugins-xgboost @ git+https://github.com/flyteorg/flytekit.git@3d1f5877ceed720da94a2fc505df1e3ae0781eeb#subdirectory=plugins/flytekit-xgboost
    # via -r requirements.in
grpcio==1.41.1
    # via flytekit
idna==3.3
    # via requests
importlib-metadata==4.8.2
    # via keyring
joblib==1.1.0
    # via scikit-learn
keyring==23.2.1
    # via flytekit
kiwisolver==1.3.2
    # via matplotlib
marshmallow==3.14.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
    #   marshmallow-jsonschema
marshmallow-enum==1.5.1
    # via dataclasses-json
marshmallow-jsonschema==0.13.0
    # via flytekit
matplotlib==3.4.3
    # via -r ../../../common/requirements-common.in
mypy-extensions==0.4.3
    # via typing-inspect
natsort==8.0.0
    # via flytekit
numpy==1.21.4
    # via
    #   matplotlib
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
    #   xgboost
pandas==1.3.4
    # via flytekit
pathspec==0.9.0
    # via scantree
pillow==8.4.0
    # via matplotlib
protobuf==3.19.1
    # via
    #   flyteidl
    #   flytekit
py==1.11.0
    # via retry
pyarrow==3.0.0
    # via flytekit
pyparsing==3.0.5
    # via matplotlib
python-dateutil==2.8.1
    # via
    #   croniter
    #   flytekit
    #   matplotlib
    #   pandas
python-json-logger==2.0.2
    # via flytekit
pytimeparse==1.1.8
    # via flytekit
pytz==2018.4
    # via
    #   flytekit
    #   pandas
regex==2021.11.2
    # via docker-image-py
requests==2.26.0
    # via
    #   flytekit
    #   responses
responses==0.15.0
    # via flytekit
retry==0.9.2
    # via flytekit
scantree==0.0.1
    # via dirhash
scikit-learn==1.0.1
    # via flytekitplugins-xgboost
scipy==1.7.2
    # via
    #   scikit-learn
    #   xgboost
six==1.16.0
    # via
    #   flytekit
    #   grpcio
    #   python-dateutil
    #   responses
    #   scantree
sortedcontainers==2.4.0
    # via flytekit
statsd==3.3.0
    # via flytekit
threadpoolctl==3.0.0
    # via scikit-learn
typing-extensions==3.10.0.2
    # via typing-inspect
typing-inspect==0.7.1
    # via dataclasses-json
urllib3==1.26.7
    # via
    #   flytekit
    #   requests
    #   responses
wheel==0.37.0
    # via
    #   -r ../../../common/requirements-common.in
    #   flytekit
wrapt==1.13.3
    # via
    #   deprecated
    #   flytekit
xgboost==1.5.0
    # via flytekitplugins-xgboost
zipp==3.6.0
    # via importlib-metadata
